moves client adapter type interfaces to client adapters folder
